Added g_config global variable which will get assigned in gameLoad. Added a config argument to gameLoad.
Removed references to the global variable 'scale', since it's being replaced by g_config.scale.

Commented out some lines which dealt with the (still unseen) UI, since they referenced the now-deleted scale, and were causing the game to crash. This is bad, but the UI is going to be revamped anyway.

Moved collision behaviour out of addEntityX functions and into registerBehaviours, which the addEntityX functions now call.

Added addEntityTrigger, which populates the g_entityTriggers table.

Added g_cameras and addCamera, which populates it. Due to how the game treats cameras, I have opted to omit the word 'entity' from this table and function.

In cases where objects aren't given names in the Tiled file, we now add the object's ID to its dummy name to give it some means of identification.