Changed a few things; the most important of which are as follows:
1) Changed up how events are loaded in and updated the tiled level accordingly.

A tiled object's collision behaviour properties are now in the key=value format, eg:
timer=0 target=door_startroom type=doorswitch_open

This means I can keep per-event logic outside of gameMain.lua and leave it for g_collisionBehaviours to handle. Basically, it makes things way easier for me.

2) Added the g_pixelLocked flag (which I thought I'd already committed, but clearly hadn't)

When this flag is on, it ensures that everything will be drawn aligned to the grid defined by the (scaled) pixels in the game's graphics. Essentially, this means (for example) that a pixel representing the player character's foot must be directly on top of a pixel of the floor graphic; the blue foot pixel cannot be partly in a grey floor pixel and partly in a white one.

This also applies to camera movement; camera positions can't move in increments of less than one scaled pixel per frame. This makes movement look weird and jolty, so I might end up removing it.

3) Fixed "require" paths as per the external libraries' new locations.

4) Added jSplitKV, which splits a string into a table of key-value pairs. Probably not implemented efficiently, but whatever.