Moved the loop which calls onCollide on g_entityBlocks elements beyond the loop which calls update on the same objects. This ensures events will be queued up and ready to be processed before the next time the g_entityBlocks update loop is hit.
Renamed g_showBoxes to g_debugDraw. Debug draw mode will now draw the collision shapes on top of the game world (as opposed to not drawing the game world at all).

Added new shaders, debugShaderWorld and debugShaderDebug, with source objects debugShaderWorldSource and debugShaderDebugSource in shaders.lua. When the debug overlay is activated, the former will darken the world slightly, and the latter will make the debug overlay slightly transparent. The intended result is that both can be seen without either interfering with the other too much.